% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ex.data}
\alias{ex.data}
\title{Example Data File}
\format{\preformatted{'data.frame':	21 obs. of  8 variables:
 $ x            : num  0 1 2 3 4 5 6 7 8 9 ...
 $ overall      : num  0.5 1.51 2.56 3.59 4.61 5.62 6.65 7.69 8.7 9.71 ...
 $ g1           : num  0.47 1.72 3.1 3.99 5.2 5.98 7.03 8.01 8.97 9.61 ...
 $ g2           : num  0.61 1.52 2.46 3.54 4.71 5.6 6.64 7.7 8.81 9.72 ...
 $ g3           : num  0.53 1.61 2.89 3.62 4.81 5.23 6.64 7.21 8.75 9.81 ...
 $ g4           : num  0.12 0.98 1.46 1.99 2.47 3.42 4.58 5.12 5.99 6.78 ...
 $ g5           : num  1.54 2.72 3.46 4.21 5.94 ...
 $ relative_freq: num  0.001 0.001 0.001 0.01 0.02 0.04 0.052 0.065 0.09 0.12 ...
}}
\usage{
ex.data
}
\value{
An example data file (class: data.frame) of six sets of equated scores conditioned on unequated (i.e., x) scores
}
\description{
An example data set that has column 1 as \eqn{x}{x} (i.e., unequated scale), column 2 as the equated scores based on the overall population, column 3 as the equated scores based on subpopulation 1, column 4 as the equated scores based on subpopulation 2, column 5 as the equated scores based on subpopulation 3, column 6 as the equated scores based on subpopulation 4, column 7 as the equated scores based on subpopulation 5, and column 8 as the frequency of examinees in the overall population scoring at each level of \eqn{x}{x}. 
}
\author{
Anne Corinne Huggins-Manley
}
\keyword{datasets}

